fix of meta duplication, when meta value is empty or null (#259)
* fix of meta duplication, when meta value is empty or null #258
refactoring of set_meta function

* fix of meta duplication, when meta value is empty or null #258
refactoring of set_meta function

* fix undefined variable notice.

* PHPCS and other formatting fixes

* Don't store the response of `update_post_meta()` (yet)

We should definitely report back any errors in storing meta; however we need to do it for both update and add and it needs to store into an identifiable array since it's inside a loop. Let's pick that up for 1.4 for sure.

* Cast `$meta_values` to an array first

Older versions of Distributor might not return an array for everything, this doesn't guarantee back-compat but at least it shouldn't explode anymore.