Start of some clarifications on passwords and connection name
I am not loving these yet so I've grouped them in an intermediate commit. Passwords need a pointed reminder that Application Passwords is the way to go and to somehow incorporate the magic link thing. Connection name is really an arbitrary label - "Enter external connection name" made it sound like if you didn't match something precisely it wouldn't work. It needs to be clearer that it's just an arbitrary label so you can tell things apart.