Fix and simplify `is_using_gutenberg()`
`has_blocks()` does a `get_post()` if you don't pass a string, which doesn't work for external posts (which presumably is what you're pulling). Better to pass a string.

Also simplifies the rest of the function down to just reproducing the check from `has_blocks()` if you're not in WP 5.0+.