Explicitly order pull list table by ID DESC
This is being done in order to support truncating the post include/exclude argument that is passed, as that can only sanely be ordered numerically. We may be able to support ASC at some point, but no need to explore that right now. It would not really be correct to keep the column named "Date" at that point but it's possibly close enough for most cases.

This removes the sortable column headers which do not appear to have ever worked.