Run NetworkSiteConnection::update_syndicated on correct hook when using Gutenberg
Add a check to the NetworkSiteConnection::update_syndicated method to see if the current post is saved through Gutenberg. If so add a new action that runs the same method but on the rest_after_insert_{$post_type} hook to fix an issue where updates to terms/meta etc where not beeing syndicated. Fixes #464