Run composer install because lol git
The georgestephanis/application-passwords package contains the `.git` directory, which means the actual files don't get committed to this repository, just a reference. So it's a non-directory in the eyes of a fresh checkout, and composer install needs to be run to populate the directory again.

We need to fix this the rest of the way and not have these vendor directories committed and referenced directly in the next release.