Ensure rest_route valid for plain permalink requests.
A URL with the `rest_route` set but empty will show the home page of the site rather than the index of the REST API endpoint. This causes the validation of these requests to fail.

By ensuring the request to the REST API index always finishes with a trailing slash, this ensures the site receives the expected JSON response data.